test 3 mechine, all normal

+++ autoTest.sh
@@ -9,24 +9,10 @@
 ## change the java version command
 ## update-alternatives --config java
 
-#################################################################
-## $1 : virtual mechine or real mechine (v/r)
-## $2 : run android_x86.iso or raw in qemu (iso/raw), for automatic installation iso have not been implemented, you can not input raw  
-## $3 : if $1=v, $3 stand for location of raw or iso
-##      if $1=r, $3 stand for location of android_x86 system (which sda?)
-## $4 : cts command
-
-## example:
-## virtual mechine: sudo ./autoTest.sh v raw ../rawiso/android_x86.raw "--plan CTS --disable-reboot"
-## virtual mechine: sudo ./autoTest.sh v raw ~/work/cts/android-x86-6.0.raw "-p android.acceleration --disable-reboot"
-## real mechine: sudo ./autoTest.sh r 192.168.2.16 /dev/sda5 "--plan CTS --disable-reboot"
-## real mechine: sudo ./autoTest.sh r 192.168.2.16 /dev/sda5 "-p android.acceleration --disable-reboot"
-#################################################################
-
 ##########################################################################################################################
 ## $1 : virtual mechine or real mechine (v/r)
 ## $2 : run android_x86(run) or install android_x86.iso(install)
-## $3 : ip of client linux system, if you test local android_x86, use localhost of 127.0.0.1
+## $3 : ip of client linux system, if you test local android_x86, use localhost or 127.0.0.1
 ## $4 : path of disk(/dev/sda40) or virtual disk(../rawiso/android_x86.raw)
 ## $5 : cts command("-p android.acceleration --disable-reboot")
 ## if test mechine has been installed android_x86, on the other world, $2=run, this 5 parameter above is enough
@@ -34,6 +20,8 @@
 ## but if we need to install a new android_x86, we should also specify the path of android_x86.iso, so $6 is needed  
 ## $6 : path of android_x86.iso(../rawiso/android_x86.iso)
 #########################################################################################################################
+## eg: ./autoTest.sh r install 192.168.2.16 /dev/sda40 "-p android.acceleration --disable--reboot" android_x86.iso
+## eg: ./autoTest.sh v install localhost /media/aquan/000D204000041550/research/android-x86-6.0.raw "-p android.acceleration --disable-reboot" android_x86.iso
 
 r_v=$1
 run_install=$2
@@ -43,6 +31,37 @@ cts_cmd=$5
 
 ip_linux_host=`/sbin/ifconfig -a|grep inet|grep -v 127.0.0.1|grep -v inet6|awk '{print $2}'|tr -d "addr:"`
 
+function EditBoot()
+{
+	if [ ! -d "android_disk" ]; then
+		mkdir  android_disk
+	fi
+	mount -o loop,offset=32256 $disk_path android_disk;
+    ########################################
+    ## modify init.sh
+    line2bottom=`tail android_disk/android*/system/etc/init.sh -n 2 |head -n 1`
+    sed '$d' -i ./android_disk/android*/system/etc/init.sh
+	sed '$d' -i ./android_disk/android*/system/etc/init.sh
+
+	#echo \$ip | nc -q 0 $ip_linux_host 5556
+	if [ "$line2bottom" == "" ]; then
+		echo "ip=\`getprop | grep ipaddress\`
+		ip=\${ip##*\[}
+		ip=\${ip%]*}
+		nc $ip_linux_host 5556 << EOF
+    	\$ip
+EOF
+		return 0" >> ./android_disk/android*/system/etc/init.sh
+	else
+    	sed '$d' -i ./android_disk/android*/system/etc/init.sh
+    	sed '$d' -i ./android_disk/android*/system/etc/init.sh
+    	echo "nc $ip_linux_host 5556 << EOF
+    	\$ip
+EOF
+    	return 0" >> ./android_disk/android*/system/etc/init.sh
+	fi
+	umount android_disk;
+}
 
 ## according to where it's virtual mechine(qemu) or real mechine, we should change the network model
 if [ "$r_v" == "v" ]; then
@@ -55,41 +74,9 @@ if [ "$r_v" == "v" ]; then
 		fi
 		#cp iso_loc ./android_x86.iso
 		./fastboot_vir.sh $disk_path flashall $iso_loc;
-	fi
+		EditBoot
 
-	#sleep 40
-	if [ "$run_install" == "run" ]|| [ "$run_install" == "install" ];then
-
-		if [ ! -d "android_disk" ]; then
-			mkdir  android_disk
-		fi
-		mount -o loop,offset=32256 $disk_path android_disk;
-        ########################################
-        ## modify init.sh
-        line2bottom=`tail android_disk/android*/system/etc/init.sh -n 2 |head -n 1`
-        sed '$d' -i ./android_disk/android*/system/etc/init.sh
-		sed '$d' -i ./android_disk/android*/system/etc/init.sh
-
-		#echo \$ip | nc -q 0 $ip_linux_host 5556
-		if [ "$line2bottom" == "" ]; then
-			echo "ip=\`getprop | grep ipaddress\`
-			ip=\${ip##*\[}
-			ip=\${ip%]*}
-			nc $ip_linux_host 5556 << EOF
-    		\$ip
-EOF
-			return 0" >> ./android_disk/android*/system/etc/init.sh
-		else
-    		sed '$d' -i ./android_disk/android*/system/etc/init.sh
-    		sed '$d' -i ./android_disk/android*/system/etc/init.sh
-    		echo "nc $ip_linux_host 5556 << EOF
-    		\$ip
-EOF
-    		return 0" >> ./android_disk/android*/system/etc/init.sh
-		fi
-        
-		umount android_disk;
-
+		## install CtsDeviceAdmin.apk and active the device adminstrators, this setting will take effect after reboot 
 		qemu-system-x86_64 -m 2G --enable-kvm -net nic -net user,hostfwd=tcp::5558-:5555 $disk_path &
 		{
 			ip_android_v=`nc -lp 5556`
@@ -97,13 +84,34 @@ EOF
 			## so it's just a symbol that android-x86 is running 
 	        echo 'waiting for android boot !!!!!'  
             sleep 60
+			adb connect localhost:5558
+			## install CtsDeviceAdmin.apk
+            echo 'install CtsDeviceAdmin.apk!!!!!'
+            adb install ../android-cts/repository/testcases/CtsDeviceAdmin.apk
+            adb push device_policies.xml data/system/device_policies.xml
+            adb shell poweroff
+		}
+
+	fi
+
+	#sleep 40
+	if [ "$run_install" == "run" ]|| [ "$run_install" == "install" ];then
+
+		EditBoot
+		qemu-system-x86_64 -m 2G --enable-kvm -net nic -net user,hostfwd=tcp::5558-:5555 $disk_path &
+		{
+			ip_android_v=`nc -lp 5556`
+			## waiting for a message from android-x86, this ip address is useful in real mechine test, but in virtural mechine ,we adopt nat address mapping ,
+			## so it's just a symbol that android-x86 is running 
+	        echo 'waiting for android boot !!!!!'  
+            sleep 60
+
             ## gui haven't been loaded completely for android_x86-5.1 
             echo 'testing'
 			adb connect localhost:5558
 			echo "exit" | ../android-cts/tools/cts-tradefed run cts $cts_cmd 
             adb shell poweroff
 		}
-
 	fi
 elif [ "$r_v" == "r" ];then
 
@@ -114,7 +122,6 @@ elif [ "$r_v" == "r" ];then
 		ssh root@${ip_linux_client} "~/script/reboot.sh $disk_path $ip_linux_host";
 
 		##ssh root@${ip_linux_client}
-
 		## return to linux_host command
 		ip_android_r=`nc -lp 5556`
 		echo $ip_android_r
